requirements.txt
# Core dependencies
google-genai>=0.3.0
openpyxl>=3.1.0
python-docx>=1.1.0
python-pptx>=0.6.23
streamlit>=1.28.0

# PDF dependencies
PyPDF2>=3.0.0
reportlab>=4.0.0

# Testing dependencies
hypothesis>=6.92.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Environment management
python-dotenv>=1.0.0
